+++ packages/core-js/actual/set/intersection.js
@@ -1,6 +1,5 @@
 'use strict';
-require('../../modules/es.set');
+var parent = require('../../stable/set/intersection');
 require('../../modules/esnext.set.intersection.v2');
-var entryUnbind = require('../../internals/entry-unbind');
 
-module.exports = entryUnbind('Set', 'intersection');
+module.exports = parent;