+++ packages/core-js/actual/set/is-disjoint-from.js
@@ -1,6 +1,5 @@
 'use strict';
-require('../../modules/es.set');
+var parent = require('../../stable/set/is-disjoint-from');
 require('../../modules/esnext.set.is-disjoint-from.v2');
-var entryUnbind = require('../../internals/entry-unbind');
 
-module.exports = entryUnbind('Set', 'isDisjointFrom');
+module.exports = parent;