+++ packages/core-js/actual/set/union.js
@@ -1,6 +1,5 @@
 'use strict';
-require('../../modules/es.set');
+var parent = require('../../stable/set/union');
 require('../../modules/esnext.set.union.v2');
-var entryUnbind = require('../../internals/entry-unbind');
 
-module.exports = entryUnbind('Set', 'union');
+module.exports = parent;