+++ packages/core-js/stage/3.js
@@ -8,7 +8,6 @@ require('../proposals/explicit-resource-management');
 require('../proposals/float16');
 require('../proposals/iterator-helpers-stage-3-2');
 require('../proposals/json-parse-with-source');
-require('../proposals/set-methods-v2');
 // TODO: Obsolete versions, remove from `core-js@4`
 require('../proposals/array-grouping-stage-3');
 require('../proposals/array-grouping-stage-3-2');